README.md: Update actuator description.
This patch updates the description of the actuator, with the original
goal being to drop the machine-api-operator reference.  There's not
dependency on that operator specifically.

This patch also updates the description of the Machine - BareMetalHost
mapping, as that has now been implemented.

The operator-sdk link was moved into the baremetal-operator
section, as that's where it's most relevant.

Finally, add some `...` usage for literals, like BareMetalHost.

Signed-off-by: Russell Bryant <rbryant@redhat.com>