README.md: Expand on bare metal management description.
The high level architecture diagram includes a "Bare Metal Management
Pods" box, but this was not described in its own section like other
boxes in the diagram.  Add a new section for it and include a pointer
to the "use-ironic" design document that has more information.

While we're at it, expand `use-ironic` to further highlight the design
goal that Ironic is not the source of truth, and we must be able to
rebuild its state at any time.