Propose applying to the CNCF Sandbox
metal3-io started with very little formalized governance.  Now that
we're a year into the project and usage and collaboration is taking
off, it seems like a good time to consider next steps on project
governance.

The specific proposal covered here is to apply to the CNCF as a
sandbox project.