Create CODE_OF_CONDUCT.md
A code of conduct is required prior to being accepted as a CNCF
Sandbox project.

This template is the one github recommends as a template, and is in
use by at least one other CNCF Sandbox project (Kubevirt).  A key
detail is where reports are to be sent, which is set to be myself and
maelk.  That way it's not limited to a single person or employer.