describe baremetal-operator migration to kubebuilder
We have recently encountered several situations where we wanted to
take advantage of kubernetes API features that are not supported by
the operator-sdk.  Migrating from operator-sdk to kubebuilder will
unblock several efforts, and it will be enough work on its own that
documenting the change separately seems prudent. This proposal
describes the work we will need to do to complete that migration, with
references to a proof-of-concept implementation.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>