Propose deploy ramdisk Image build API and integration
Enable a custom image containing the ironic-python-agent to be built for
each BareMetalHost, instead of requiring all hosts to use the same agent
image. The interface to the image builder will be defined by a new
RamDiskImageRequest CR in metal³, with users able to substitute their
own implementations of image building (in most cases this is expected to
involve light customisation of an existing image) by creating a
controller. The baremetal-operator will create and use
RamDiskImageRequests as required when configured to do so.

This allows users to customise the ramdisk for each Host to contain
network settings specific to that host.