HostData for host inspection data
Instead of writing hardware details into the status of
BareMetalHost, write it to a HostData Custom Resource.
Baremetalhost and HostData will be parent and child objects
respectively, and this relationship is controlled via
ownerReferences chain.

Having quite big inspection data is problematic when pivoting
BareMetalHost objects to a target cluster. Storing the data in a
CR will avoid the issue regardless of the inspection data
size.