Add separate OWNERS file for design docs
The approvers list for design docs is intentionally kept quite small,
but we would like the freedom to add many more reviewers and approvers
for documentation in general. By adding this OWNERS file with the
no_parent_owners option set, we disconnect the design docs directory
from the rest of the repo.