Add docker-build target
This adds a docker-build make target for building the container image.
The Dockerfile is also expanded to build the releasetags preprocessor in
a build container instead of copying from the host.

The mdbook version is taken from the Makfile as an argument. The rust
version is bumped since the old version could not build the newer mdbook
version. (The versions in the Dockerfile and Makefile were different.)

Signed-off-by: Lennart Jern <lennart.jern@est.tech>