User guide: copy provisioning states information
This information is copied more or less as it is from the BMO source
code, including the diagram. I'm not exactly fond of how the information
is presented, but it's the best I can do now (and I strongly believe we
need at least something about the states in our documentation).

The only original part is the introduction and the mention of
operationalStatus/errorType.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>