Docs: Fix broken link at processes/reviewer-permissions-migration.md
- Corrects a broken link in `processes/reviewer-permissions-migration.md`.
The link [a process for approval permission](../maintainers)
used an invalid relative path as the destination link.

That link has been updated to point to the maintainers `README.md`
within the community repository at
https://github.com/metal3-io/community/blob/main/maintainers/README.md.

- Update 1st Paragraph of Implementation Details per Feedback.
Rewords the first paragraph of `Implementation Details/Notes/Contraints`
to mention the `OWNERS_ALIASES` and `OWNERS` files relation in repos.

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>